Moved html files into separate folders for organisation. Set up socketio and javascript functions to allow the host to affect the players screens
Set up the player side of the application.
Players can input a session code to join a session in progress displaying a message to wait for others to join
When the host starts the game, the ability to input question answers is given to the user.
If a player joins late the host will be able to "ping" everyone, this ensures the players are answering the most up to date question and that the waiting screen is not being displayed, this is in anticipation of an issue where refreshing will load the player into the waiting screen and will only change on the next question or a player joining late and having to wait for the next question to begin.